Use webfactory/phumbor instead of 99designs/phumbor
The `99designs/phumbor` library has been unmaintained for a long time. In particular, there is a long-standing issue regarding application of `rawurlencode()` to image URLs (https://github.com/99designs/phumbor/issues/20) which has not been addressed. Also the (direct question whether the code is still maintained or not)[https://github.com/99designs/phumbor/issues/30] has not been answered for almost a year.

In order to address the URL encoding issue, I forked the original repo:

* https://github.com/webfactory/phumbor/
* https://packagist.org/packages/webfactory/phumbor

With this change I would like to suggest switching this bundle over to the forked library.